requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.2
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
